+++ cypress/integration/contract/contract.spec.js
@@ -1,3 +1,5 @@
+export const WAIT_TIME = 3000
+
 describe('Contract', () => {
   Cypress.on('uncaught:exception', (error, runnable) => {
     console.error(error)
@@ -7,6 +9,7 @@ describe('Contract', () => {
   beforeEach(() => {
     cy.visit('/contract')
     cy.fixture('contract.json').as('data')
+    cy.fixture('holderChangePersonaldata.json').as('personaldata')
   })
 
   describe('Contract with selfconsumption', function () {
@@ -74,6 +77,9 @@ describe('Contract', () => {
     })
 
     it('3.0TD no incremental powers', function () {
+      if (!Cypress.env('FEATURE_FLAGS').is30ContractEnabled) {
+        this.skip()
+      }
       const moreThan15Kw = true
       const powers = [
         this.data.power,
@@ -88,6 +94,9 @@ describe('Contract', () => {
     })
 
     it('Contract with 3.0TD', function () {
+      if (!Cypress.env('FEATURE_FLAGS').is30ContractEnabled) {
+        this.skip()
+      }
       const moreThan15Kw = true
       const powers = [
         this.data.power,
@@ -151,6 +160,9 @@ describe('Contract', () => {
     })
 
     it('Contract with 3.0TD', function () {
+      if (!Cypress.env('FEATURE_FLAGS').is30ContractEnabled) {
+        this.skip()
+      }
       const moreThan15Kw = true
 
       cy.chooseMoreOrLessThan15Kw(moreThan15Kw)
@@ -193,6 +205,9 @@ describe('Contract', () => {
     })
 
     it('Contract 3.0A less than 15kW', function () {
+      if (!Cypress.env('FEATURE_FLAGS').is30ContractEnabled) {
+        this.skip()
+      }
       const moreThan15Kw = true
 
       cy.chooseMoreOrLessThan15Kw(moreThan15Kw)
@@ -247,8 +262,8 @@ describe('Contract', () => {
 
     cy.get('#cups')
       .clear()
-      .type(this.data.supplyPoint.invalidCups)
-      .should('have.value', this.data.supplyPoint.invalidCups)
+      .type(this.personaldata.invalidCups)
+      .should('have.value', this.personaldata.invalidCups)
       .blur()
     cy.contains('Actualmente este CUPS está en un proceso')
 
@@ -263,7 +278,7 @@ describe('Contract', () => {
       .type(this.data.supplyPoint.cups)
       .should('have.value', this.data.supplyPoint.cups)
 
-    cy.wait(2000)
+    cy.wait(WAIT_TIME)
     cy.get(`[data-value="${this.data.supplyPoint.hasService}"]`).click()
 
     cy.get('[data-cy=next]').click()
@@ -315,7 +330,7 @@ describe('Contract', () => {
       .type(this.data.holder.vat)
       .should('have.value', this.data.holder.vat)
 
-    cy.wait(2000)
+    cy.wait(WAIT_TIME)
     cy.get(`[data-value="${this.data.holder.previousHolder}"]`).click()
 
     cy.get('[data-cy=next]').click()
@@ -444,7 +459,7 @@ describe('Contract', () => {
         .type(this.data.juridicMember.vat)
         .should('have.value', this.data.juridicMember.vat)
 
-      cy.wait(2000)
+      cy.wait(WAIT_TIME)
       cy.get(`[data-value="${this.data.juridicHolder.previousHolder}"]`).click()
 
       cy.get('[data-cy=next]').click()
@@ -454,7 +469,7 @@ describe('Contract', () => {
       cy.get('[name="holder.vat"]')
         .type(this.data.juridicHolder.vat)
         .should('have.value', this.data.juridicHolder.vat)
-      cy.wait(2000)
+      cy.wait(WAIT_TIME)
       cy.get(`[data-value="${this.data.juridicHolder.previousHolder}"]`).click()
 
       cy.get('[data-cy=next]').click()