+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 - Gurb: validation pages for non-members
 - Added feature flag: isGurbEnabled
-
+- Gurb: Add new member form
+- Add member and godmother/godfather validation pages
 
 ## 2.4.1 2024-11-11
 - Added message for autonomous (Modify Contract form)